reorganize payments files (#10235)
