+++ test/api/v3/integration/payments/amazon/GET-payments_amazon_subscribe_cancel.test.js
@@ -3,7 +3,7 @@ import {
   generateGroup,
   translate as t,
 } from '../../../../../helpers/api-integration/v3';
-import amzLib from '../../../../../../website/server/libs/amazonPayments';
+import amzLib from '../../../../../../website/server/libs/payments/amazon';
 
 describe('payments : amazon #subscribeCancel', () => {
   let endpoint = '/amazon/subscribe/cancel?noRedirect=true';
