+++ test/api/v3/integration/payments/amazon/POST-payments_amazon_checkout.test.js
@@ -1,7 +1,7 @@
 import {
   generateUser,
 } from '../../../../../helpers/api-integration/v3';
-import amzLib from '../../../../../../website/server/libs/amazonPayments';
+import amzLib from '../../../../../../website/server/libs/payments/amazon';
 
 describe('payments - amazon - #checkout', () => {
   let endpoint = '/amazon/checkout';
