+++ test/api/v3/integration/payments/apple/GET-payments_apple_cancelSubscribe.js
@@ -1,5 +1,5 @@
 import {generateUser} from '../../../../../helpers/api-integration/v3';
-import applePayments from '../../../../../../website/server/libs/applePayments';
+import applePayments from '../../../../../../website/server/libs/payments/apple';
 
 describe('payments : apple #cancelSubscribe', () => {
   let endpoint = '/iap/ios/subscribe/cancel?noRedirect=true';
