+++ test/api/v3/integration/payments/apple/POST-payments_apple_verifyiap.js
@@ -1,5 +1,5 @@
 import {generateUser} from '../../../../../helpers/api-integration/v3';
-import applePayments from '../../../../../../website/server/libs/applePayments';
+import applePayments from '../../../../../../website/server/libs/payments/apple';
 
 describe('payments : apple #verify', () => {
   let endpoint = '/iap/ios/verify';
