+++ test/api/v3/integration/payments/google/GET-payments_google_cancelSubscribe.js
@@ -1,5 +1,5 @@
 import {generateUser} from '../../../../../helpers/api-integration/v3';
-import googlePayments from '../../../../../../website/server/libs/googlePayments';
+import googlePayments from '../../../../../../website/server/libs/payments/google';
 
 describe('payments : google #cancelSubscribe', () => {
   let endpoint = '/iap/android/subscribe/cancel?noRedirect=true';
