+++ test/api/v3/integration/payments/google/POST-payments_google_subscribe.test.js
@@ -1,5 +1,5 @@
 import {generateUser, translate as t} from '../../../../../helpers/api-integration/v3';
-import googlePayments from '../../../../../../website/server/libs/googlePayments';
+import googlePayments from '../../../../../../website/server/libs/payments/google';
 
 describe('payments : google #subscribe', () => {
   let endpoint = '/iap/android/subscribe';
