+++ test/api/v3/integration/payments/paypal/GET-payments_paypal_checkout.test.js
@@ -1,7 +1,7 @@
 import {
   generateUser,
 } from '../../../../../helpers/api-integration/v3';
-import paypalPayments from '../../../../../../website/server/libs/paypalPayments';
+import paypalPayments from '../../../../../../website/server/libs/payments/paypal';
 
 describe('payments : paypal #checkout', () => {
   let endpoint = '/paypal/checkout';
