+++ test/api/v3/integration/payments/stripe/POST-payments_stripe_subscribe_edit.test.js
@@ -3,7 +3,7 @@ import {
   generateGroup,
   translate as t,
 } from '../../../../../helpers/api-integration/v3';
-import stripePayments from '../../../../../../website/server/libs/stripePayments';
+import stripePayments from '../../../../../../website/server/libs/payments/stripe';
 
 describe('payments - stripe - #subscribeEdit', () => {
   let endpoint = '/stripe/subscribe/edit';
