+++ test/api/v3/unit/libs/payments/stripe/edit-subscription.test.js
@@ -4,7 +4,7 @@ import {
   generateGroup,
 } from '../../../../../../helpers/api-unit.helper.js';
 import { model as User } from '../../../../../../../website/server/models/user';
-import stripePayments from '../../../../../../../website/server/libs/stripePayments';
+import stripePayments from '../../../../../../../website/server/libs/payments/stripe';
 import common from '../../../../../../../website/common';
 
 const i18n = common.i18n;
