+++ test/api/v3/unit/libs/payments/stripe/handle-webhook.test.js
@@ -4,8 +4,8 @@ import {
   generateGroup,
 } from '../../../../../../helpers/api-unit.helper.js';
 import { model as User } from '../../../../../../../website/server/models/user';
-import stripePayments from '../../../../../../../website/server/libs/stripePayments';
-import payments from '../../../../../../../website/server/libs/payments';
+import stripePayments from '../../../../../../../website/server/libs/payments/stripe';
+import payments from '../../../../../../../website/server/libs/payments/payments';
 import common from '../../../../../../../website/common';
 import logger from '../../../../../../../website/server/libs/logger';
 import { v4 as uuid } from 'uuid';
