+++ website/server/controllers/top-level/payments/iap.js
@@ -5,8 +5,8 @@ import {
 import {
   BadRequest,
 } from '../../../libs/errors';
-import googlePayments from '../../../libs/googlePayments';
-import applePayments from '../../../libs/applePayments';
+import googlePayments from '../../../libs/payments/google';
+import applePayments from '../../../libs/payments/apple';
 
 let api = {};
